+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.3
 
-* TODO: Describe initial release.
+What's new in this version
+
+* A new example has been added for this pinger state management 
+* Description api has been added for the new builder widget PingBuilder which will help to
+  update the widget directly without set state